+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - TT-1828 - Increasing size of address preview field in registration journey
 - TT-1199 - Fixing accessibility issues
 - TT-2025 - Update success page content
+- No ticket - Removed old feature flagged items
 
 ### Fixed bugs:
 - TT-1817 - Use correct input types for number fields
@@ -19,6 +20,7 @@
 - TT-1727 - Not prompt collaborators to change business type
 - TT-2096 - Fix layout problem on enrolment of individual that already has sso profile
 - TT-1708 - link user to ingress url on success page
+- TT-2095 - Added "remove from business profile" feature for non-admins
 - TT-1855 - Fix incorrect company type label on enrolment
 
 ## [2019.10.30](https://github.com/uktrade/directory-sso-profile/releases/tag/2019.10.30)